requirements.txt
openai
ddgs
jinja2
python-dotenv
